Barbies.Bi

Functor

A bifunctor is simultaneously a FunctorT and a FunctorB.

btmap :: (FunctorB (b f), FunctorT b) => (forall a. f a -> f' a) -> (forall a. g a -> g' a) -> b f g -> b f' g'

Map over both arguments at the same time.

btmap1 :: (FunctorB (b f), FunctorT b) => (forall a. f a -> g a) -> b f f -> b g g

A version of btmap specialized to a single argument.

Traversable

A traversable bifunctor is simultaneously a TraversableT and a TraversableB.

bttraverse :: (TraversableB (b f), TraversableT b, Monad t) => (forall a. f a -> t (f' a)) -> (forall a. g a -> t (g' a)) -> b f g -> t (b f' g')

Traverse over both arguments, first over f, then over g..

bttraverse1 :: (TraversableB (b f), TraversableT b, Monad t) => (forall a. f a -> t (g a)) -> b f f -> t (b g g)

A version of bttraverse specialized to a single argument.

bttraverse_ :: (TraversableB (b f), TraversableT b, Monad e) => (forall a. f a -> e c) -> (forall a. g a -> e d) -> b f g -> e ()

Map each element to an action, evaluate these actions from left to right and ignore the results.

btfoldMap :: (TraversableB (b f), TraversableT b, Monoid m) => (forall a. f a -> m) -> (forall a. g a -> m) -> b f g -> m

Map each element to a monoid, and combine the results.

Applicative

If t is an ApplicativeT, the type of tpure shows that its second argument must be a phantom-type, so there are really no interesting types that are both ApplicativeT and ApplicativeB. However, we can sometimes reconstruct a bi-applicative from an ApplicativeB and a FunctorT.

btpure :: (ApplicativeB (b Unit), FunctorT b) => (forall a. f a) -> (forall a. g a) -> b f g

Conceptually, this is like simultaneously using bpure and tpure.

btpure1 :: (ApplicativeB (b Unit), FunctorT b) => (forall a. f a) -> b f f

A version of btpure specialized to a single argument.

btprod :: (ApplicativeB (b (Alt (Product f f'))), FunctorT b, Alternative f, Alternative f') => b f g -> b f' g' -> b (f `Product` f') (g `Product` g')

Simultaneous product on both arguments.

Wrappers

newtype Flip b l r

Convert a FunctorB into a FunctorT and vice-versa.

Constructors

Flip

Fields

runFlip :: b r l