Barbies.Internal

Functor

gbmapDefault :: CanDeriveFunctorB b f g => (forall a. f a -> g a) -> b f -> b g

Default implementation of bmap based on Generic.

class GFunctor (n :: Nat) f g repbf repbg where

Methods

gmap :: Proxy n -> (forall a. f a -> g a) -> repbf x -> repbg x

type CanDeriveFunctorB b f g = (GenericP 0 (b f), GenericP 0 (b g), GFunctor 0 f g (RepP 0 (b f)) (RepP 0 (b g)))

CanDeriveFunctorB B f g is in practice a predicate about B only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (B f).

B f can contain fields of type b f as long as there exists a FunctorB b instance. In particular, recursive usages of B f are allowed.

B f can also contain usages of b f under a Functor h. For example, one could use Maybe (B f) when defining B f.

type CanDeriveFunctorT t f g x = (GenericP 1 (t f x), GenericP 1 (t g x), GFunctor 1 f g (RepP 1 (t f x)) (RepP 1 (t g x)))

CanDeriveFunctorT T f g x is in practice a predicate about T only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (T f).

T f x can contain fields of type t f y as long as there exists a FunctorT t instance. In particular, recursive usages of T f y are allowed.

T f x can also contain usages of t f y under a Functor h. For example, one could use Maybe (T f y) when defining T f y.

Traversable

gbtraverseDefault :: forall b f g e. (Applicative e, CanDeriveTraversableB b f g) => (forall a. f a -> e (g a)) -> b f -> e (b g)

Default implementation of btraverse based on Generic.

class GTraversable n f g repbf repbg where

Methods

gtraverse :: Applicative t => Proxy n -> (forall a. f a -> t (g a)) -> repbf x -> t (repbg x)

type CanDeriveTraversableB b f g = (GenericP 0 (b f), GenericP 0 (b g), GTraversable 0 f g (RepP 0 (b f)) (RepP 0 (b g)))

CanDeriveTraversableB B f g is in practice a predicate about B only. It is analogous to CanDeriveFunctorB, so it essentially requires the following to hold, for any arbitrary f:

There is an instance of Generic (B f).

B f can contain fields of type b f as long as there exists a TraversableB b instance. In particular, recursive usages of B f are allowed.

B f can also contain usages of b f under a Traversable h. For example, one could use Maybe (B f) when defining B f.

type CanDeriveTraversableT t f g x = (GenericP 1 (t f x), GenericP 1 (t g x), GTraversable 1 f g (RepP 1 (t f x)) (RepP 1 (t g x)))

CanDeriveTraversableT T f g x is in practice a predicate about T only. It is analogous to CanDeriveFunctorT, so it essentially requires the following to hold, for any arbitrary f:

There is an instance of Generic (T f x).

T f x can contain fields of type t f x as long as there exists a TraversableT t instance. In particular, recursive usages of T f x are allowed.

T f x can also contain usages of t f x under a Traversable h. For example, one could use Maybe (T f x) when defining T f x.

Distributive

gbdistributeDefault :: CanDeriveDistributiveB b f g => Functor f => f (b g) -> b (Compose f g)

Default implementation of bdistribute based on Generic.

class Functor f => GDistributive (n :: Nat) f repbg repbfg where

Methods

gdistribute :: Proxy n -> f (repbg x) -> repbfg x

type CanDeriveDistributiveB b f g = (GenericP 0 (b g), GenericP 0 (b (Compose f g)), GDistributive 0 f (RepP 0 (b g)) (RepP 0 (b (Compose f g))))

CanDeriveDistributiveB B f g is in practice a predicate about B only. Intuitively, it says the the following holds for any arbitrary f:

There is an instance of Generic (B f).

(B f) has only one constructor, and doesn't contain "naked" fields (that is, not covered by f).

B f can contain fields of type b f as long as there exists a DistributiveB b instance. In particular, recursive usages of B f are allowed.

B f can also contain usages of b f under a Distributive h. For example, one could use a -> (B f) as a field of B f.

type CanDeriveDistributiveT (t :: (Type -> Type) -> i -> Type) f g x = (GenericP 1 (t g x), GenericP 1 (t (Compose f g) x), GDistributive 1 f (RepP 1 (t g x)) (RepP 1 (t (Compose f g) x)))

CanDeriveDistributiveT T f g x is in practice a predicate about T only. Intuitively, it says the the following holds for any arbitrary f:

There is an instance of Generic (B f x).

(B f x) has only one constructor, and doesn't contain "naked" fields (that is, not covered by f). In particular, x needs to occur under f.

B f x can contain fields of type b f y as long as there exists a DistributiveT b instance. In particular, recursive usages of B f x are allowed.

B f x can also contain usages of b f y under a Distributive h. For example, one could use a -> (B f x) as a field of B f x.

Applicative

gbpureDefault :: forall b f. CanDeriveApplicativeB b f f => (forall a. f a) -> b f

gbprodDefault :: forall b f g. CanDeriveApplicativeB b f g => b f -> b g -> b (f `Product` g)

Default implementation of bprod based on Generic.

class GApplicative n (f :: k -> Type) (g :: k -> Type) repbf repbg repbfg where

Methods

gprod :: Proxy n -> Proxy f -> Proxy g -> repbf x -> repbg x -> repbfg x

gpure :: (f ~ g, repbf ~ repbg) => Proxy n -> Proxy f -> Proxy repbf -> Proxy repbfg -> (forall a. f a) -> repbf x

type CanDeriveApplicativeB b f g = (GenericP 0 (b f), GenericP 0 (b g), GenericP 0 (b (f `Product` g)), GApplicative 0 f g (RepP 0 (b f)) (RepP 0 (b g)) (RepP 0 (b (f `Product` g))))

CanDeriveApplicativeB B f g is in practice a predicate about B only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (B f).

B has only one constructor (that is, it is not a sum-type).

Every field of B f is either a monoid, or of the form f a, for some type a.

type CanDeriveApplicativeT t f g x = (GenericP 1 (t f x), GenericP 1 (t g x), GenericP 1 (t (f `Product` g) x), GApplicative 1 f g (RepP 1 (t f x)) (RepP 1 (t g x)) (RepP 1 (t (f `Product` g) x)))

CanDeriveApplicativeT T f g x is in practice a predicate about T only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (T f).

T has only one constructor (that is, it is not a sum-type).

Every field of T f x is either a monoid, or of the form f a, for some type a.

Constraints

gbaddDictsDefault :: forall b c f. (CanDeriveConstraintsB c b f, AllB c b) => b f -> b (Dict c `Product` f)

Default implementation of baddDicts based on Generic.

class GConstraints n c f repbx repbf repbdf where

Methods

gaddDicts :: GAll n c repbx => repbf x -> repbdf x

type CanDeriveConstraintsB c b f = (GenericN (b f), GenericN (b (Dict c `Product` f)), AllB c b ~ GAll 0 c (GAllRepB b), GConstraints 0 c f (GAllRepB b) (RepN (b f)) (RepN (b (Dict c `Product` f))))

CanDeriveConstraintsB B f g is in practice a predicate about B only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (B f).

B f can contain fields of type b f as long as there exists a ConstraintsB b instance. In particular, recursive usages of B f are allowed.

type CanDeriveConstraintsT c t f x = (GenericN (t f x), GenericN (t (Dict c `Product` f) x), AllT c t ~ GAll 1 c (GAllRepT t), GConstraints 1 c f (GAllRepT t) (RepN (t f x)) (RepN (t (Dict c `Product` f) x)))

CanDeriveConstraintsT T f g x is in practice a predicate about T only. Intuitively, it says that the following holds, for any arbitrary f and x:

There is an instance of Generic (T f x).

T f can contain fields of type t f x as long as there exists a ConstraintsT t instance. In particular, recursive usages of T f x are allowed.

type family GAll (n :: Nat) (c :: k -> Constraint) (repbf :: Type -> Type) :: Constraint

type GAllRepB b = TagSelf0 b

The representation used for the generic computation of the AllB c b constraints.

type GAllRepT t = TagSelf1 t

The representation used for the generic computation of the AllT c t constraints. .

data X a

data family Y :: k

data Self (p :: Type) (a :: Type) (x :: Type)

data Other (p :: Type) (a :: Type) (x :: Type)

type family SelfOrOther (b :: k) (b' :: k) :: Type -> Type -> Type -> Type where ...

Equations

SelfOrOther b b = Self

SelfOrOther b b' = Other

type TagSelf0 b = TagSelf0' (Indexed b 1) (RepN (b X))

We use the type-families to generically compute AllB c b. Intuitively, if b' f' occurs inside b f, then we should just add AllB b' c to AllB b c. The problem is that if b is a recursive type, and b' is b, then ghc will choke and blow the stack (instead of computing a fixpoint).

So, we would like to behave differently when b = b' and add () instead of AllB b c to break the recursion. Our trick will be to use a type family to inspect Rep (b X), for an arbitrary X, and distinguish recursive usages from non-recursive ones, tagging them with different types, so we can distinguish them in the instances.

type family TagSelf0' (b :: kf -> Type) (repbf :: Type -> Type) :: Type -> Type where ...

Equations

TagSelf0' b (M1 mt m s) = M1 mt m (TagSelf0' b s)

TagSelf0' b (l :+: r) = TagSelf0' b l :+: TagSelf0' b r

TagSelf0' b (l :*: r) = TagSelf0' b l :*: TagSelf0' b r

TagSelf0' (b :: kf -> Type) (Rec ((b' :: kf -> Type) f) ((b'' :: kf -> Type) g)) = SelfOrOther b b' (b' f) (b'' g)

TagSelf0' b (Rec x y) = Rec x y

TagSelf0' b U1 = U1

TagSelf0' b V1 = V1

type TagSelf1 b = TagSelf1' (Indexed b 2) (Zip (Rep (Indexed (b X) 1 Y)) (Rep (b X Y)))

We use the type-families to generically compute AllT c b. Intuitively, if t' f' x' occurs inside t f x, then we should just add AllT t' c to AllT t c. The problem is that if t is a recursive type, and t' is t, then ghc will choke and blow the stack (instead of computing a fixpoint).

So, we would like to behave differently when t = t' and add () instead of AllT t c to break the recursion. Our trick will be to use a type family to inspect Rep (t X Y), for arbitrary X and Y and distinguish recursive usages from non-recursive ones, tagging them with different types, so we can distinguish them in the instances.

type family TagSelf1' (b :: kf -> kg -> Type) (repbf :: Type -> Type) :: Type -> Type where ...

Equations

TagSelf1' b (M1 mt m s) = M1 mt m (TagSelf1' b s)

TagSelf1' b (l :+: r) = TagSelf1' b l :+: TagSelf1' b r

TagSelf1' b (l :*: r) = TagSelf1' b l :*: TagSelf1' b r

TagSelf1' (b :: kf -> kg -> Type) (Rec ((b' :: kf -> kg -> Type) fl fr) ((b'' :: kf -> kg -> Type) gl gr)) = SelfOrOther b b' (b' fl gr) (b'' gl gr)

TagSelf1' b (Rec x y) = Rec x y

TagSelf1' b U1 = U1

TagSelf1' b V1 = V1

Bare values

gbcoverDefault :: CanDeriveBareB b => b Bare Identity -> b Covered Identity

Default implementation of bstrip based on Generic.

gbstripDefault :: CanDeriveBareB b => b Covered Identity -> b Bare Identity

Default implementation of bstrip based on Generic.

class GBare (n :: Nat) repbi repbb where

Methods

gstrip :: Proxy n -> repbi x -> repbb x

gcover :: Proxy n -> repbb x -> repbi x

type CanDeriveBareB b = (GenericP 0 (b Bare Identity), GenericP 0 (b Covered Identity), GBare 0 (RepP 0 (b Covered Identity)) (RepP 0 (b Bare Identity)))

All types that admit a generic FunctorB instance, and have all their occurrences of f under a Wear admit a generic BareB instance.

Generic derivation support

data family Param (n :: Nat) (a :: k) :: k

type family Indexed (t :: k) (i :: Nat) :: k where ...

Equations

Indexed (t a) i = Indexed t (i + 1) (Param i a)

Indexed t _ = t

type family FilterIndex (n :: Nat) (t :: k) :: k where ...

Equations

FilterIndex n (t (Param n a)) = FilterIndex n t (Param n a)

FilterIndex n (t (Param _ a)) = FilterIndex n t a

FilterIndex _ t = t

type family Zip (a :: Type -> Type) (b :: Type -> Type) :: Type -> Type where ...

Equations

Zip (M1 mt m s) (M1 mt m t) = M1 mt m (Zip s t)

Zip (l :+: r) (l' :+: r') = Zip l l' :+: Zip r r'

Zip (l :*: r) (l' :*: r') = Zip l l' :*: Zip r r'

Zip (Rec0 p) (Rec0 a) = Rec p a

Zip U1 U1 = U1

Zip V1 V1 = V1

newtype Rec (p :: Type) a x

Constructors

Rec

Fields

unRec :: K1 R a x

class (Coercible (Rep a) (RepN a), Generic a) => GenericN (a :: Type) where

Associated Types

type RepN (a :: Type) :: Type -> Type

type RepN a = Zip (Rep (Indexed a 0)) (Rep a)

Methods

toN :: RepN a x -> a

fromN :: a -> RepN a x

class (Coercible (Rep a) (RepP n a), Generic a) => GenericP (n :: Nat) (a :: Type) where

Associated Types

type RepP n a :: Type -> Type

type RepP n a = Zip (Rep (FilterIndex n (Indexed a 0))) (Rep a)

Methods

toP :: Proxy n -> RepP n a x -> a

fromP :: Proxy n -> a -> RepP n a x

module GHC.Generics