Data.Barbie.Internal

Functor

gbmapDefault :: CanDeriveFunctorB b f g => (forall a. f a -> g a) -> b f -> b g

Default implementation of bmap based on Generic.

class GFunctorB f g repbf repbg where

Methods

gbmap :: (forall a. f a -> g a) -> repbf x -> repbg x

type CanDeriveFunctorB b f g = (GenericN (b f), GenericN (b g), GFunctorB f g (RepN (b f)) (RepN (b g)))

CanDeriveFunctorB B f g is in practice a predicate about B only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (B f).

B f can contain fields of type b f as long as there exists a FunctorB b instance. In particular, recursive usages of B f are allowed.

B f can also contain usages of b f under a Functor h. For example, one could use Maybe (B f) when defining B f.

Traversable

gbtraverseDefault :: forall b f g t. (Applicative t, CanDeriveTraversableB b f g) => (forall a. f a -> t (g a)) -> b f -> t (b g)

Default implementation of btraverse based on Generic.

class GTraversableB f g repbf repbg where

Methods

gbtraverse :: Applicative t => (forall a. f a -> t (g a)) -> repbf x -> t (repbg x)

type CanDeriveTraversableB b f g = (GenericN (b f), GenericN (b g), GTraversableB f g (RepN (b f)) (RepN (b g)))

CanDeriveTraversableB B f g is in practice a predicate about B only. It is analogous to CanDeriveFunctorB, so it essentially requires the following to hold, for any arbitrary f:

There is an instance of Generic (B f).

B f can contain fields of type b f as long as there exists a TraversableB b instance. In particular, recursive usages of B f are allowed.

B f can also contain usages of b f under a Traversable h. For example, one could use Maybe (B f) when defining B f.

Product

gbuniqDefault :: forall b f. CanDeriveProductB b f f => (forall a. f a) -> b f

gbprodDefault :: forall b f g. CanDeriveProductB b f g => b f -> b g -> b (f `Product` g)

Default implementation of bprod based on Generic.

class GProductB (f :: k -> *) (g :: k -> *) repbf repbg repbfg where

Methods

gbprod :: repbf x -> repbg x -> repbfg x

gbuniq :: (forall a. f a) -> repbf x

type CanDeriveProductB b f g = (GenericN (b f), GenericN (b g), GenericN (b (f `Product` g)), GProductB f g (RepN (b f)) (RepN (b g)) (RepN (b (f `Product` g))))

CanDeriveProductB B f g is in practice a predicate about B only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (B f).

B has only one constructor (that is, it is not a sum-type).

Every field of B f is of the form f a, for some type a. In other words, B has no "hidden" structure.

Constraints

gbaddDictsDefault :: forall b c f. (CanDeriveConstraintsB c b f, AllB c b) => b f -> b (Dict c `Product` f)

Default implementation of baddDicts based on Generic.

class GAllBC repbx => GConstraintsB c (f :: k -> *) repbx repbf repbdf where

Methods

gbaddDicts :: GAllB c repbx => repbf x -> repbdf x

type CanDeriveConstraintsB c b f = (GenericN (b f), GenericN (b (Dict c `Product` f)), AllB c b ~ GAllB c (GAllBRep b), GConstraintsB c f (GAllBRep b) (RepN (b f)) (RepN (b (Dict c `Product` f))))

CanDeriveConstraintsB B f g is in practice a predicate about B only. Intuitively, it says that the following holds, for any arbitrary f:

There is an instance of Generic (B f).

B f can contain fields of type b f as long as there exists a ConstraintsB b instance. In particular, recursive usages of B f are allowed.

class GAllBC (repbf :: * -> *)

Associated Types

type GAllB (c :: k -> Constraint) repbf :: Constraint

type GAllBRep b = TagSelf b (RepN (b X))

The representation used for the generic computation of the AllB c b constraints. Here X is an arbitrary constant since the actual argument to b is irrelevant.

data X a

type family TagSelf (b :: (k -> *) -> *) (repbf :: * -> *) :: * -> * where ...

We use type-families to generically compute AllB c b. Intuitively, if b' f occurs inside b f, then we should just add AllB b' c to AllB b c. The problem is that if b is a recursive type, and b' is b, then ghc will choke and blow the stack (instead of computing a fixpoint).

So, we would like to behave differently when b = b' and add () instead of `AllB b f` to break the recursion. Our trick will be to use a type family to inspect RepN (b f) and distinguish recursive usages from non-recursive ones, tagging them with different types, so we can distinguish them in the instances.

Equations

TagSelf b (M1 mt m s) = M1 mt m (TagSelf b s)

TagSelf b (l :+: r) = TagSelf b l :+: TagSelf b r

TagSelf b (l :*: r) = TagSelf b l :*: TagSelf b r

TagSelf b (Rec (b f) (b g)) = Rec (Self b f) (b g)

TagSelf (b :: (k -> *) -> *) (Rec (b' f) ((b'' :: (k -> *) -> *) g)) = Rec (Other b' f) (b'' g)

TagSelf b (Rec p a) = Rec p a

TagSelf b U1 = U1

TagSelf b V1 = V1

data Self (b :: (k -> *) -> *) (f :: k -> *)

data Other (b :: (k -> *) -> *) (f :: k -> *)

Proof

gbdictsDefault :: forall b c. (CanDeriveProductBC c b, AllB c b) => b (Dict c)

Default implementation of bproof based on Generic.

class GProductBC c repbx repbd where

Methods

gbdicts :: GAllB c repbx => repbd x

type CanDeriveProductBC c b = (GenericN (b (Dict c)), AllB c b ~ GAllB c (GAllBRep b), GProductBC c (GAllBRep b) (RepN (b (Dict c))))

Every type that admits a generic instance of ProductB and ConstraintsB, has a generic instance of ProductBC as well.

Bare values

gbcoverDefault :: CanDeriveBareB b => b Bare Identity -> b Covered Identity

Default implementation of bstrip based on Generic.

gbstripDefault :: CanDeriveBareB b => b Covered Identity -> b Bare Identity

Default implementation of bstrip based on Generic.

class GBareB repbi repbb where

Methods

gbstrip :: repbi x -> repbb x

gbcover :: repbb x -> repbi x

type CanDeriveBareB b = (GenericN (b Bare Identity), GenericN (b Covered Identity), GBareB (RepN (b Covered Identity)) (RepN (b Bare Identity)))

All types that admit a generic FunctorB' instance, and have all their occurrences of f under a Wear admit a generic BareB instance.

Generic derivation support

class (Coercible (Rep a) (RepN a), Generic a) => GenericN (a :: Type)

Minimal complete definition

toN, fromN

newtype Rec (p :: Type) a x

Constructors

Rec

Fields

unRec :: K1 R a x

type family RepN (a :: Type) :: Type -> Type