Data.Functor.Prod

Description

Deprecated: The module is no longer part of the main api and will be removed

Generalize the standard two-functor Product to the product of n-functors. Intuitively, this means:

Product f g a ~~ (f a, g a)

Prod '[]        a ~~  Const () a
Prod '[f]       a ~~ (f a)
Prod '[f, g]    a ~~ (f a, g a)
Prod '[f, g, h] a ~~ (f a, g a, h a)
    ⋮

n-tuples of functors.

data Prod :: [k -> Type] -> k -> Type where

Product of n functors.

Constructors

Unit :: Prod '[] a

Cons :: f a -> Prod fs a -> Prod (f ': fs) a

zeroTuple :: Prod '[] a

The unit of the product.

oneTuple :: f a -> Prod '[f] a

Lift a functor to a 1-tuple.

fromProduct :: Product f g a -> Prod '[f, g] a

Conversion from a standard Product

toProduct :: Prod '[f, g] a -> Product f g a

Conversion to a standard Product

Flat product of functor products

prod :: Prod ls a -> Prod rs a -> Prod (ls ++ rs) a

Flat product of products.

Lifting functions

uncurryn :: Curried (Prod fs a -> r a) -> Prod fs a -> r a

Like uncurry but using Prod instead of pairs. Can be thought of as a family of functions:

uncurryn :: r a -> Prod '[] a
uncurryn :: (f a -> r a) -> Prod '[f] a
uncurryn :: (f a -> g a -> r a) -> Prod '[f, g] a
uncurryn :: (f a -> g a -> h a -> r a) -> Prod '[f, g, h] a
        ⋮

Type-level helpers

type family l ++ r :: [k] where ...

Type-level, poly-kinded, list-concatenation.

Equations

'[] ++ ys = ys

(x ': xs) ++ ys = x ': (xs ++ ys)

type family Curried t where ...

Prod '[f, g, h] a -> r is the type of the uncurried form of a function f a -> g a -> h a -> r. Curried moves from the former to the later. E.g.

Curried (Prod '[]  a    -> r) = r a
Curried (Prod '[f] a    -> r) = f a -> r a
Curried (Prod '[f, g] a -> r) = f a -> g a -> r a

Equations

Curried (Prod '[] a -> r a) = r a

Curried (Prod (f ': fs) a -> r a) = f a -> Curried (Prod fs a -> r a)